Extract RefCountedObserverMap
This is more broadly useful for flexible reactive serverside publication,
and I want to make use of it for publishing chat notifications (which may
also reference a Hunt and a Puzzle).
